fix: restructure directory
