+++ app/api/profile/update/route.ts
@@ -21,9 +21,7 @@ interface UpdateProfileData {
 
 export const POST = async (req: Request) => {
   if (req.method !== "POST") {
-    return new NextResponse(`Method ${req.method} Not Allowed`, {
-      status: 405,
-    });
+    return new NextResponse(`Method ${req.method} Not Allowed`, { status: 405 });
   }
 
   const session = await getServerSession({ req, ...authOptions });
