+++ app/api/topup/route.ts
@@ -10,37 +10,20 @@ const prisma = new PrismaClient();
   return this.toString();
 };
 
-const transactionSchema = z.object({
-  amount: z.string().min(4, { message: "Amount must be at least 4 digit" }),
-  cardNumber: z.string().min(3, { message: "Card number must be at least 3 characters long" }),
-  expirationDate: z
-    .string()
-    .date()
-    .refine((date) => new Date(date) > new Date(), { message: "Expiration date must be a future date" }),
-  cvc: z.string().min(3, { message: "CVC must be at least 3 characters long" }),
-  cardHolderName: z.string().min(3, { message: "Card holder name must be at least 3 characters long" }),
-});
-
 export const POST = async (req: Request) => {
+  if (req.method !== "POST") {
+    return new NextResponse(`Method ${req.method} Not Allowed`, { status: 405 });
+  }
+
+  const session = await getServerSession({ req, ...authOptions });
+  if (!session) {
+    return NextResponse.json({ message: "Not authenticated" }, { status: 401 });
+  }
+
   try {
-    const session = await getServerSession({ req, ...authOptions });
-    if (!session) {
-      return NextResponse.json({ message: "Not authenticated" }, { status: 401 });
-    }
-
     const body = await req.json();
-    const parsedBody = transactionSchema.safeParse(body);
-    if (!parsedBody.success) {
-      return NextResponse.json(
-        {
-          message: "Invalid input",
-          errors: parsedBody.error.errors,
-        },
-        { status: 400 }
-      );
-    }
 
-    const { amount, cardNumber, expirationDate, cvc, cardHolderName } = parsedBody.data;
+    const { amount, cardNumber, expirationDate, cvc, cardHolderName } = body;
 
     const newData: any = {
       userId: session.user.id,
